move Toleration to Pod object
