+++ client/src/main/scala/skuber/json/package.scala
@@ -7,7 +7,6 @@ import java.time.format._
 import org.apache.commons.codec.binary.Base64
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
-import skuber.Toleration.{EqualToleration, ExistsToleration, TolerationEffect}
 import skuber._
 
 /**
@@ -232,11 +231,11 @@ package object format {
   implicit val secCtxtFormat: Format[Security.Context] = Json.format[Security.Context]
 
 
-  implicit val tolerationEffectFmt: Format[Toleration.TolerationEffect] = new Format[Toleration.TolerationEffect] {
+  implicit val tolerationEffectFmt: Format[Pod.TolerationEffect] = new Format[Pod.TolerationEffect] {
 
-    import Toleration.Effects._
+    import Pod.TolerationEffect._
 
-    override def reads(json: JsValue): JsResult[Toleration.TolerationEffect] = json match {
+    override def reads(json: JsValue): JsResult[Pod.TolerationEffect] = json match {
       case JsString(value) => value match {
         case NoSchedule.name => JsSuccess(NoSchedule)
         case PreferNoSchedule.name => JsSuccess(PreferNoSchedule)
@@ -245,46 +244,44 @@ package object format {
       case _ => JsError(s"Toleration effect should be a string")
     }
 
-    override def writes(effect: Toleration.TolerationEffect): JsValue = effect match {
+    override def writes(effect: Pod.TolerationEffect): JsValue = effect match {
       case NoSchedule => JsString(NoSchedule.name)
       case PreferNoSchedule => JsString(PreferNoSchedule.name)
     }
   }
 
-  implicit val tolerationFmt: Format[Toleration] = new Format[Toleration] {
+  implicit val tolerationFmt: Format[Pod.Toleration] = new Format[Pod.Toleration] {
 
-    override def reads(json: JsValue): JsResult[Toleration] = json match {
+    override def reads(json: JsValue): JsResult[Pod.Toleration] = json match {
       case JsObject(fields) if fields.contains("operator") =>
 
         val key = fields("key").as[String]
-        val effect: Option[TolerationEffect] = fields.get("effect").flatMap{
+        val effect: Option[Pod.TolerationEffect] = fields.get("effect").flatMap{
           case JsNull => None
-          case e @ _ => Some(e.as[TolerationEffect])
+          case e @ _ => Some(e.as[Pod.TolerationEffect])
         }
 
         fields("operator") match {
           case JsString("Equal") =>
             val value = fields("value").as[String]
-            JsSuccess(EqualToleration(key, value, effect))
-          case JsString("Exists") => JsSuccess(ExistsToleration(key, effect))
+            JsSuccess(Pod.EqualToleration(key, value, effect))
+          case JsString("Exists") => JsSuccess(Pod.ExistsToleration(key, effect))
           case operator => JsError(s"Unknown operator '$operator'")
         }
 
       case _ => JsError(s"Unknown toleration")
     }
 
-    //import play.api.libs.json._
+    override def writes(toleration: Pod.Toleration): JsValue = toleration match {
 
-    override def writes(toleration: Toleration): JsValue = toleration match {
-
-      case EqualToleration(key, value, effect) => Json.obj(
+      case Pod.EqualToleration(key, value, effect) => Json.obj(
         "key" -> key,
         "value" -> value,
         "operator" -> "Equal",
         "effect" -> Json.toJson(effect)
       )
 
-      case ExistsToleration(key, effect) => Json.obj(
+      case Pod.ExistsToleration(key, effect) => Json.obj(
         "key" -> key,
         "operator" -> "Exists",
         "effect" -> Json.toJson(effect)
@@ -689,7 +686,7 @@ package object format {
       (JsPath \ "hostNetwork").formatMaybeEmptyBoolean() and
       (JsPath \ "imagePullSecrets").formatMaybeEmptyList[LocalObjectReference] and
       (JsPath \ "affinity").formatNullable[Pod.Affinity] and
-      (JsPath \ "tolerations").formatMaybeEmptyList[Toleration]
+      (JsPath \ "tolerations").formatMaybeEmptyList[Pod.Toleration]
     )(Pod.Spec.apply _, unlift(Pod.Spec.unapply))
     
   implicit val podTemplSpecFormat: Format[Pod.Template.Spec] = Json.format[Pod.Template.Spec]
